+++ src/commands/hardis/package/version/create.ts
@@ -123,11 +123,11 @@ export default class PackageVersionCreate extends SfdxCommand {
     // Create package version
     uxLog(this, c.cyan(`Generating new package version for ${c.green(pckgDirectory.package)}...`));
     const createCommand =
-      "sfdx force:package:version:create" +
+      "sf package version create" +
       ` --package "${pckgDirectory.package}"` +
-      (this.installKey ? ` --installationkey "${this.installKey}"` : " --installationkeybypass") +
-      " --codecoverage" +
-      " -w 60";
+      (this.installKey ? ` --installation-key "${this.installKey}"` : " --installation-key-bypass") +
+      " --code-coverage" +
+      " --wait 60";
     const createResult = await execSfdxJson(createCommand, this, {
       fail: true,
       output: true,
@@ -139,7 +139,7 @@ export default class PackageVersionCreate extends SfdxCommand {
     if (this.deleteAfter) {
       // Delete package version
       uxLog(this, c.cyan(`Delete new package version ${c.green(latestVersion)} of package ${c.green(pckgDirectory.package)}...`));
-      const deleteVersionCommand = "sfdx force:package:version:delete --noprompt -p " + latestVersion;
+      const deleteVersionCommand = "sf package version delete --no-prompt --package " + latestVersion;
       const deleteVersionResult = await execSfdxJson(deleteVersionCommand, this, {
         fail: true,
         output: true,